Refactoring variable code to be cleaner
